fix: ui changes and list page in other projects
